bootloader: move bootloader.h to include/
Do this to avoid having to use relative includes.
