+++ subsys/bootloader/include/linker.ld
@@ -16,7 +16,7 @@
 
 #include <autoconf.h>
 /* TODO: remove file when MULTI_IMAGE */
-#include <../bootloader.h>
+#include <bootloader.h>
 #include <linker/sections.h>
 #include <generated_dts_board.h>
